+++ spec/features/projects/users_can_view_a_list_of_assigned_projects_spec.rb
@@ -74,7 +74,7 @@ RSpec.feature "Viewing assigned projects" do
 
       expect(page).to have_content(I18n.t("project.user.in_progress.title"))
 
-      within(".projects-list") do
+      within(".govuk-table") do
         expect(page).to have_content(in_progress_assigned_project.urn)
         expect(page).not_to have_content(completed_assigned_project.urn)
 
@@ -83,9 +83,6 @@ RSpec.feature "Viewing assigned projects" do
 
         expect(page).not_to have_content(completed_other_user_project.urn)
         expect(page).not_to have_content(completed_unassigned_project.urn)
-
-        expect(page.find("h2.govuk-heading-l:first-of-type").text).to eq((Date.today + 1.month).at_beginning_of_month.strftime("%B %Y"))
-        expect(page.find("h2.govuk-heading-l:last-of-type").text).to eq((Date.today + 1.year).at_beginning_of_month.strftime("%B %Y"))
       end
     end