+++ spec/features/your_projects/users_can_view_their_projects_spec.rb
@@ -16,16 +16,10 @@ RSpec.feature "Users can view their projects" do
     expect(page).to have_content(project.establishment.name)
     expect(page).to have_content(project.urn)
     expect(page).to have_content(project.incoming_trust.name)
-    expect(page).to have_content("provisional")
-    expect(page).to have_content("Not yet")
-    expect(page).to have_content("Region")
+    expect(page).to have_content(I18n.t("none"))
   end
 
   scenario "they can see transfers assigned to them" do
-    # TODO:  we have to disable bullet for this test as we eager load the TasksData for conversions, which
-    # use the all_conditions_met value, but transfers do not use it. Once transfers has the concept of
-    # all conditions met - this can be removed
-    Bullet.enable = false
     project = create(:transfer_project, assigned_to: user)
 
     visit in_progress_your_projects_path
@@ -34,9 +28,6 @@ RSpec.feature "Users can view their projects" do
     expect(page).to have_content(project.urn)
     expect(page).to have_content(project.incoming_trust.name)
     expect(page).to have_content(project.outgoing_trust.name)
-    expect(page).to have_content("provisional")
-    expect(page).to have_content("Region")
-    Bullet.enable = true
   end
 
   context "when they are NOT in the Regional casework services team" do